Cleaning up for ram freeing
Cleaned unused functions as to not unnecessarily use ram

Cleaned message parsing to be used better overall less ram usage

Moved Azimuth Range to its own header file

Changed buffer size to not be so huge

Previous problem was that it was getting overloaded and did not have enough ram to run and stopped after first few instances in setup

RECOMMENDED: At arduino uno limits we should upgrade to something with more ram soon

+++ Arduino/New/Ino_codes/Arduino_Ground_Station/AzimuthFull.h
@@ -27,8 +27,8 @@ static const int   AZ_MAX_ITER = 200;
 
 /* =================== FUNCTION =================== */
 static inline AzimuthResult azmuth_Range(
-    const double coordinate1[2],
-    const double coordinate2[2],
+    const float coordinate1[2],
+    const float coordinate2[2],
     float altitude1,
     float altitude2)
 {
